backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database (SQLite)
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-decouple==3.8

# CORS
fastapi-cors==0.0.6

# Validation
pydantic==2.5.0
pydantic-settings==2.0.3

# Email (for password reset)
fastapi-mail==1.4.1

# File handling
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0